Analysis of the trajectory
Analysis of the trajectory of close contacts in the spread of new coronary pneumonia( COVID19).This function describes how to analyze and output its train, bus, and ship peers based on the time window corresponding to the past travel trajectory of the infected person's mobile phone number based on the mobile phone number recorded on the base station of the telecommunications operator Mobile phone number, and how to provide disease control centers, emergency management departments, community managers with data query, SMS broadcast notification, Internet notification technology system. It can be widely used for peer inquiry, crowd gathering places such as shopping mall epidemic spread tracking. The main problems encountered in the multi-level analysis are how to solve the problem of calculating the coexistence time of the infected person and the close contacts in different time periods during the cumulative time period, the cumulative deduplication problem found at each level, and the minimum time to eliminate accidental passing problem. Through this function, combined with user basic data, base station or wifi device location and location correspondence data, you can output the corresponding close contact situation in the specified time period according to the requirements of the analysis level. This method is very valuable for the analysis of the unclear propagation path, and can effectively avoid the epidemic prevention and control problems caused by the unknown transmission route due to inaccurate memory and concealment of historical traces.